+++ src/apps/mesoscale-explorer/ui/states.tsx
@@ -17,7 +17,6 @@ import { PluginContext } from '../../../mol-plugin/context';
 import { StateAction, StateObjectRef, StateTransform } from '../../../mol-state';
 import { Task } from '../../../mol-task';
 import { Color } from '../../../mol-util/color/color';
-import { ColorNames } from '../../../mol-util/color/names';
 import { getFileInfo } from '../../../mol-util/file-info';
 import { ParamDefinition as PD } from '../../../mol-util/param-definition';
 import { createCellpackHierarchy } from '../data/cellpack/preset';
@@ -27,7 +26,7 @@ function adjustPluginProps(ctx: PluginContext) {
     ctx.managers.interactivity.setProps({ granularity: 'chain' });
     ctx.canvas3d?.setProps({
         multiSample: { mode: 'off' },
-        cameraClipping: { far: false },
+        cameraClipping: { far: false, minNear: 50 },
         renderer: {
             colorMarker: true,
             highlightColor: Color(0xffffff),
@@ -59,6 +58,7 @@ function adjustPluginProps(ctx: PluginContext) {
                     distanceFactor: 10,
                     blurKernelSize: 11,
                     resolutionScale: 1,
+                    color: Color(0x000000),
                 }
             },
             shadow: {
@@ -75,7 +75,7 @@ function adjustPluginProps(ctx: PluginContext) {
                 params: {
                     scale: 1,
                     threshold: 0.33,
-                    color: ColorNames.black,
+                    color: Color(0x000000),
                     includeTransparent: true,
                 }
             }